+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Bug Fixes
 - PR #209 remove deprecated cudf methods- `to_gpu_matrix`, `add_column` and groupby parameter `method`
 - PR #212 remove redundant docs folders and files, removed bloated notebooks
+- PR #214 fix map_style in choropleths, and fix custom_binning param issue in core_aggregate charts
 
 # cuXfilter 0.16.0 (Date TBD)